+++ packages/snap/src/handlers/cronjob.tsx
@@ -1,4 +1,5 @@
 import type { JsonRpcRequest } from '@metamask/snaps-sdk';
+import { type Types, utils } from 'tronweb';
 
 import type { PriceApiClient } from '../clients/price-api/PriceApiClient';
 import type { SnapClient } from '../clients/snap/SnapClient';
@@ -377,8 +378,21 @@ export class CronHandler {
       return;
     }
 
-    const { preferences, scope, fromAddress, origin, scanParameters } =
-      interfaceContext;
+    const { preferences, scope, fromAddress, origin } = interfaceContext;
+    /**
+     * We know that the `transactionRawData` object is valid JSON but Tronweb returns some
+     * fields as `unknown` so we must manually cast it to the correct type.
+     */
+    const transactionRawData = interfaceContext.transactionRawData as
+      | Types.Transaction['raw_data']
+      | null;
+
+    if (!transactionRawData) {
+      this.#logger.info(
+        'Context is missing transactionRawData for scan refresh',
+      );
+      return;
+    }
 
     try {
       // Update UI to show fetching state for scan
@@ -413,12 +427,7 @@ export class CronHandler {
       try {
         scan = await this.#transactionScanService.scanTransaction({
           accountAddress: fromAddress,
-          parameters: {
-            from: scanParameters?.from ?? undefined,
-            to: scanParameters?.to ?? undefined,
-            data: scanParameters?.data ?? undefined,
-            value: scanParameters?.value ?? undefined,
-          },
+          transactionRawData,
           origin,
           scope,
           options,
@@ -552,7 +561,7 @@ export class CronHandler {
       return;
     }
 
-    const { preferences, scope, account, origin, scanParameters } =
+    const { preferences, scope, account, origin, transaction } =
       interfaceContext;
 
     // Determine what needs to be refreshed
@@ -588,15 +597,16 @@ export class CronHandler {
           options.push('validation');
         }
 
+        const transactionRawData: Types.Transaction['raw_data'] =
+          utils.deserializeTx.deserializeTransaction(
+            transaction.type,
+            transaction.rawDataHex,
+          );
+
         try {
           scan = await this.#transactionScanService.scanTransaction({
             accountAddress: account.address,
-            parameters: {
-              from: scanParameters?.from ?? undefined,
-              to: scanParameters?.to ?? undefined,
-              data: scanParameters?.data ?? undefined,
-              value: scanParameters?.value ?? undefined,
-            },
+            transactionRawData,
             origin,
             scope,
             options,